Refine copy skip detection and modularize directory executor
