+++ crates/engine/src/local_copy/executor/directory/mod.rs
@@ -1,0 +1,6 @@
+mod planner;
+mod recursive;
+mod support;
+
+pub(crate) use recursive::copy_directory_recursive;
+pub(crate) use support::{is_device, is_fifo};
