+++ crates/engine/src/local_copy/executor/directory/support.rs
@@ -1,0 +1,92 @@
+use std::cmp::Ordering;
+use std::ffi::{OsStr, OsString};
+use std::fs;
+use std::path::{Path, PathBuf};
+
+use crate::local_copy::LocalCopyError;
+
+#[derive(Debug)]
+pub(crate) struct DirectoryEntry {
+    pub(crate) file_name: OsString,
+    pub(crate) path: PathBuf,
+    pub(crate) metadata: fs::Metadata,
+}
+
+pub(crate) fn read_directory_entries_sorted(
+    path: &Path,
+) -> Result<Vec<DirectoryEntry>, LocalCopyError> {
+    let mut entries = Vec::new();
+    let read_dir = fs::read_dir(path)
+        .map_err(|error| LocalCopyError::io("read directory", path.to_path_buf(), error))?;
+
+    for entry in read_dir {
+        let entry = entry.map_err(|error| {
+            LocalCopyError::io("read directory entry", path.to_path_buf(), error)
+        })?;
+        let entry_path = entry.path();
+        let metadata = fs::symlink_metadata(&entry_path).map_err(|error| {
+            LocalCopyError::io("inspect directory entry", entry_path.to_path_buf(), error)
+        })?;
+        entries.push(DirectoryEntry {
+            file_name: entry.file_name(),
+            path: entry_path,
+            metadata,
+        });
+    }
+
+    entries.sort_by(|a, b| compare_file_names(&a.file_name, &b.file_name));
+    Ok(entries)
+}
+
+fn compare_file_names(left: &OsStr, right: &OsStr) -> Ordering {
+    #[cfg(unix)]
+    {
+        use std::os::unix::ffi::OsStrExt;
+
+        left.as_bytes().cmp(right.as_bytes())
+    }
+
+    #[cfg(windows)]
+    {
+        use std::os::windows::ffi::OsStrExt;
+
+        let left_wide: Vec<u16> = left.encode_wide().collect();
+        let right_wide: Vec<u16> = right.encode_wide().collect();
+        left_wide.cmp(&right_wide)
+    }
+
+    #[cfg(not(any(unix, windows)))]
+    {
+        left.to_string_lossy().cmp(&right.to_string_lossy())
+    }
+}
+
+pub(crate) fn is_fifo(file_type: &fs::FileType) -> bool {
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::FileTypeExt;
+
+        file_type.is_fifo()
+    }
+
+    #[cfg(not(unix))]
+    {
+        let _ = file_type;
+        false
+    }
+}
+
+pub(crate) fn is_device(file_type: &fs::FileType) -> bool {
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::FileTypeExt;
+
+        file_type.is_char_device() || file_type.is_block_device()
+    }
+
+    #[cfg(not(unix))]
+    {
+        let _ = file_type;
+        false
+    }
+}
